refactor(invest): remove retired app legacy surface
