--- frontend/invest/src/components/AppShell.tsx
@@ -1,5 +1,0 @@
-import type { ReactNode } from "react";
-
-export function AppShell({ children }: { children: ReactNode }) {
-  return <div className="app-shell">{children}</div>;
-}
