--- frontend/invest/src/components/discover/DiscoverHeader.tsx
@@ -1,11 +1,0 @@
-// frontend/invest/src/components/discover/DiscoverHeader.tsx
-export function DiscoverHeader() {
-  return (
-    <div style={{ paddingTop: 4 }}>
-      <h1 style={{ margin: 0, fontSize: 22, fontWeight: 800 }}>발견</h1>
-      <div className="subtle" style={{ marginTop: 4 }}>
-        뉴스 기반 참고 정보입니다. 매매 추천이 아닙니다.
-      </div>
-    </div>
-  );
-}
