+++ frontend/invest/src/types/filters.ts
@@ -1,0 +1,3 @@
+import type { AssetCategory } from "./invest";
+
+export type AssetCategoryKey = AssetCategory | "all";
